OCE-329 Renamed tender-rules.jsx
